Added Xee 2.2
Xee is a lightweight, fast and convenient image viewer and browser. It
is designed to be a serious tool for image viewing and management, with
a sleek and powerful interface.

Xee is useful as a more powerful replacement for Preview, or most any
other image viewer available on Mac OS X. It is very fast, and uses
less memory than most other image viewing tools. It also uses OpenGL to
display and scroll images much more smoothly than other viewers. Its
interface is highly streamlined and has easily configurable keyboard
shortcuts.