Add Sophos Anti-Virus Home Edition version 9.1p
The current production version of Sophos anti virus is 9.0.X and has a
terrible bug in it related to interfering with CORS requests in Chrome
and Firefox. They have claimed that it is remediated in 9.1, but it is
only in a preview stage right now. If you want to test it out though,
here is the formula to test it.

This formula differs from the main one in that it will automatically
install Sophos completely. In the main formula, the Sophos installer is
downloaded, but has to be run separately by the user.