Android Studio 1.0 RC4
Updating Android Studio to RC4 and removing the postflight section.

As of version RC3 (http://tools.android.com/recent) the user can export which JDK he wants Android Studio to use. The postflight section was changing the Info.plist therefore causing trouble for Android Studio to auto update. So, now that we have way to change the JDK without changing the Info.plist, we should use it.