Add Scansnap Manager V22L11+V22L16
In order to scan from the ScanSnap S300M (and some other models), you need to download the Scansnap Manager V22L11 and then install the V22L16 update.

There is a currently a problem, where uninstalling either cask will actually uninstall both, because it is impossible to just uninstall the update or the original without uninstalling the other. I am not sure what the best thing to do in this case is.

In reality you never want to install V22L11 without installing V22L16 after, but I made to formulas because this is the easiest way to achieve this and was reccomended when I asked how to install two package for one formula (caskroom/homebrew-cask#6788)