rstudio-preview: update to always pull latest
Instead of submitting an updated formula for every update, I've
modified the formula to mimic the iterm-nightly formula.

I submitted a similar pull request that used my own URL but was
told that was, understandably, an unsafe approach.

RStudio has kindly provided a link that will always point to the
latest Preview version of RStudio and I have updated the formula
to use this URL.

Thanks RStudio!