Change android-studio-canary to match mainline
The Android Studio Canary cask seems to have been last synchronized with
the changes in the mainline cask on 11 December 2014, in commit
cf066db5f26e4f2b8a994a5d1792b1ded94b6150. This commit adds all the
subsequent updates on the mainline cask, which are as follows:

 - Added name stanza.
 - Removed obsolete caveat for using Java 7 or above
 - Added zap stanza

The homepage has been changed to the homepage for the canary channel.