target: Should Only Be Used in Select Cases
See https://github.com/caskroom/homebrew-cask/blob/master/doc/cask_language_reference/stanzas/app.md#target-should-only-be-used-in-select-cases

If the developers haven’t felt the need to name the stable vs development versions differently, we also won’t do it. Our installations are to be as close to possible to manual installs.