Add GPG Suite Beta (2016.12b1) (#3155)
This release fixes the compatibility problems with MacOS 10.12 Sierra.
The announcement which includes known issues is available at
https://gpgtools.tenderapp.com/discussions/problems/49449-will-not-work-on-macosx-sierra

This Cask is a copy of the gpgtools cask from the main repo. The only
changes are the name and the appcast url. The appcast does not
currently have the beta listed. I assume that future betas will appear
correctly. I found the url by enabling the check beta versions option in
preferences.