Add beta version of Blender (v.2.80)
This cask adds the beta version (v.2.80 in our case right now) to homebrew casks. It doesn't conflict with the stable version. The download url is generated by the occurrences for the version 2.8x and will has to be updated if the version gets when the version upgrades. Used a generated download url, because there is a new version every day, which includes bug fixes. I used the installation script that is used in the original blender cask too. Removed the blenderplayer, because it doesn't exist in 2.8x yet or anymore.