Update and rename mono-mdk516.rb to mono-mdk-vs.rb
The Visual Studio channel has been updated to version 6.x for quite some time. Wondering if it would be more feasible to just rename this cask to properly denote the channel maintained by the developers or leave this one on an older, unsupported and create a new cask for the Visual Studio branch.