openemu-experimental: fix depends_on macos according to releasers (#9401)
* openemu-experimental: fix depends_on macos according to releasers

After installed, when we open the application, the following warning message is showed:

```
Your Mac is too old for this version of OpenEmu

Since version 2.1, OpenEmu requires a GPU which supports Metal. Metal is available on all Macs that support macOS 10.14 or later.

As OpenEmu does not work with your current hardware/software configuration, it will now close.
```

The current formula for openemu in homebrew-cask was used as reference (sha256 were adjusted):

https://github.com/Homebrew/homebrew-cask/blob/e927fea62652ff5d6223ce3cb491ad981b914934/Casks/openemu.rb

* Add auto_updates and conflicts_with

As pointed by @miccal:

> https://github.com/Homebrew/homebrew-cask-versions/pull/9401#pullrequestreview-458874539

* Update openemu-experimental.rb

Co-authored-by: Vítor Galvão <info@vitorgalvao.com>